:bug: 드롭다운 - 다이얼로그 트리거 방식 변경
