+++ src/components/kanban-board/index.ts
@@ -1,4 +1,4 @@
 export * from './board';
 export * from './board-sidebar';
 export * from './board-add-dialog';
-export * from './board-edit-dialog';
+export * from './board-edit-dialog-content';
