+++ src/components/kanban-header/header-command.tsx
@@ -15,7 +15,6 @@ import {
 import { cn } from '@/lib';
 import { useKanbanStore } from '@/store';
 import { BoardAddDialog } from '@/components';
-import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { type BoardId } from '@/schema';
 
@@ -27,7 +26,6 @@ const HeaderCommand = () => {
   const board = boards[currentBoardId];
 
   const boardList = Object.values(boards);
-  const router = useRouter();
 
   const onSearch = (boardId: string, term: string) => {
     // 검색어를 입력할 때마다 매칭 점수를 평가하기 위해 각 boardId에 대한 필터링이 독립적으로 실행됨
@@ -69,8 +67,8 @@ const HeaderCommand = () => {
               ))}
             </CommandGroup>
           </CommandList>
-          <BoardAddDialog onConfirm={({ id, title }) => router.push(`${id}?title=${title}`)}>
-            <Button className="rounded-none font-bold capitalize lg:hidden">add board</Button>
+          <BoardAddDialog asChild>
+            <Button className="m-2 font-bold capitalize lg:hidden">add board</Button>
           </BoardAddDialog>
         </Command>
       </PopoverContent>
